Optimize TaskConfig storage footprint
interval: u64 -> u32. Seconds-between-executions never needs more than
~136 years of range, unlike last_run (a ledger timestamp, kept u64 to
match env.ledger().timestamp()) or gas_balance (a token amount, kept
i128). blocked_by/yield_strategy stay u64 for consistency with every
other ID in the contract (task IDs, portfolio IDs, etc. are all u64).

Benchmark: added test_task_config_storage_footprint, which measures a
TaskConfig's actual XDR-serialized size (the thing that determines
ledger storage/rent, as opposed to the CPU/memory the existing
test_gas.rs benchmarks measure) and asserts it as a regression lock.
Measured 412 bytes before this change, 408 after - a 4-byte reduction
per stored TaskConfig, matching the XDR ScVal encoding difference
between U64 (4-byte type tag + 8-byte value) and U32 (4-byte type tag +
4-byte value).

Also fixed track_gas() in test_gas.rs, which computed cpu/memory cost
but discarded it without printing (stale 'println not available in
wasm tests' comment - these are native tests, not wasm) - it now prints
via std::eprintln! so cargo test -- --nocapture actually benchmarks
something.

Closes #674

+++ contract/src/lib.rs
@@ -162,7 +162,10 @@ pub struct TaskConfig {
     pub function: Symbol,
     pub args: Vec<Val>,
     pub resolver: Option<Address>,
-    pub interval: u64,
+    /// Minimum seconds between executions. `u32` is sufficient here (max
+    /// ~136 years) - unlike `last_run`, which is a ledger timestamp and must
+    /// stay `u64` to match `env.ledger().timestamp()`.
+    pub interval: u32,
     pub last_run: u64,
     pub gas_balance: i128,
     pub whitelist: Vec<Address>,
@@ -1149,7 +1152,7 @@ impl SoroTaskContract {
                 .persistent()
                 .get::<DataKey, TaskConfig>(&DataKey::Task(task_id))
             {
-                if config.is_active && now >= config.last_run + config.interval {
+                if config.is_active && now >= config.last_run + config.interval as u64 {
                     executable.push_back(ExecutableTask {
                         task_id,
                         target: config.target,
@@ -1927,7 +1930,7 @@ impl SoroTaskContract {
                 .persistent()
                 .get::<DataKey, TaskConfig>(&DataKey::Task(task_id))
             {
-                if config.is_active && now >= config.last_run + config.interval {
+                if config.is_active && now >= config.last_run + config.interval as u64 {
                     executable.push_back(ExecutableTask {
                         task_id,
                         target: config.target,
@@ -1980,7 +1983,7 @@ impl SoroTaskContract {
             panic_with_error!(env, Error::Unauthorized);
         }
 
-        if env.ledger().timestamp() < config.last_run + config.interval {
+        if env.ledger().timestamp() < config.last_run + config.interval as u64 {
             return;
         }
